+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/PortForward.java
@@ -16,6 +16,7 @@
 package io.fabric8.kubernetes.client;
 
 import java.io.Closeable;
+import java.util.Collection;
 
 /**
  * A handle that allows closing a port forward operation and checking its status.
@@ -23,5 +24,11 @@ import java.io.Closeable;
 public interface PortForward extends Closeable {
 
   boolean isAlive();
+
+  boolean errorOccurred();
+
+  Collection<Throwable> getClientThrowables();
+
+  Collection<Throwable> getServerThrowables();
 
 }