+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/PortForwarderWebsocket.java
@@ -25,6 +25,9 @@ import java.nio.channels.ReadableByteChannel;
 import java.nio.channels.ServerSocketChannel;
 import java.nio.channels.SocketChannel;
 import java.nio.channels.WritableByteChannel;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -103,6 +106,16 @@ public class PortForwarderWebsocket implements PortForwarder {
         }
 
         @Override
+        public boolean errorOccurred() {
+          for (PortForward handle : handles) {
+            if (handle.errorOccurred()) {
+              return true;
+            }
+          }
+          return false;
+        }
+
+        @Override
         public InetAddress getLocalAddress() {
           try {
             return ((InetSocketAddress) server.getLocalAddress()).getAddress();
@@ -118,6 +131,24 @@ public class PortForwarderWebsocket implements PortForwarder {
           } catch (IOException e) {
             throw new IllegalStateException("Cannot determine local address", e);
           }
+        }
+
+        @Override
+        public Collection<Throwable> getClientThrowables() {
+          Collection<Throwable> clientThrowables = new ArrayList<>();
+          for (PortForward handle : handles) {
+            clientThrowables.addAll(handle.getClientThrowables());
+          }
+          return clientThrowables;
+        }
+
+        @Override
+        public Collection<Throwable> getServerThrowables() {
+          Collection<Throwable> serverThrowables = new ArrayList<>();
+          for (PortForward handle : handles) {
+            serverThrowables.addAll(handle.getServerThrowables());
+          }
+          return serverThrowables;
         }
       };
 
@@ -150,6 +181,9 @@ public class PortForwarderWebsocket implements PortForwarder {
   @Override
   public PortForward forward(URL resourceBaseUrl, int port, final ReadableByteChannel in, final WritableByteChannel out) {
     final AtomicBoolean alive = new AtomicBoolean(true);
+    final AtomicBoolean errorOccurred = new AtomicBoolean(false);
+    final Collection<Throwable> clientThrowables = Collections.synchronizedCollection(new ArrayList<>());
+    final Collection<Throwable> serverThrowables = Collections.synchronizedCollection(new ArrayList<>());
     final String logPrefix = "FWD";
 
     Request request = new Request.Builder()
@@ -187,6 +221,7 @@ public class PortForwarderWebsocket implements PortForwarder {
 
               } catch (IOException e) {
                 if (alive.get()) {
+                  clientThrowables.add(e);
                   LOG.error("Error while writing client data");
                   closeBothWays(webSocket, 1001, "Client error");
                 }
@@ -202,6 +237,7 @@ public class PortForwarderWebsocket implements PortForwarder {
         try {
           onMessage(webSocket, ByteBuffer.wrap(text.getBytes("UTF-8")));
         } catch (IOException e) {
+          serverThrowables.add(e);
           LOG.error("Error while converting string to byte buffer", e);
           closeBothWays(webSocket, 1002, "Protocol error");
         }
@@ -221,16 +257,19 @@ public class PortForwarderWebsocket implements PortForwarder {
         }
 
         if (!buffer.hasRemaining()) {
+          errorOccurred.set(true);
           LOG.error("Received an empty message");
           closeBothWays(webSocket, 1002, "Protocol error");
         }
 
         byte channel = buffer.get();
         if (channel < 0 || channel > 1) {
+          errorOccurred.set(true);
           LOG.error("Received a wrong channel from the remote socket: {}", channel);
           closeBothWays(webSocket, 1002, "Protocol error");
         } else if (channel == 1) {
           // Error channel
+          errorOccurred.set(true);
           LOG.error("Received an error from the remote socket");
           closeForwarder();
         } else {
@@ -240,6 +279,7 @@ public class PortForwarderWebsocket implements PortForwarder {
               out.write(buffer); // channel byte already skipped
             } catch (IOException e) {
               if (alive.get()) {
+                clientThrowables.add(e);
                 LOG.error("Error while forwarding data to the client", e);
                 closeBothWays(webSocket, 1002, "Protocol error");
               }
@@ -268,6 +308,7 @@ public class PortForwarderWebsocket implements PortForwarder {
       public void onFailure(WebSocket webSocket, Throwable t, Response response) {
         LOG.debug("{}: onFailure", logPrefix);
         if (alive.get()) {
+          serverThrowables.add(t);
           LOG.error(logPrefix + ": Throwable received from websocket", t);
           closeForwarder();
         }
@@ -279,6 +320,7 @@ public class PortForwarderWebsocket implements PortForwarder {
         try {
           webSocket.close(code, message);
         } catch (Exception e) {
+          serverThrowables.add(e);
           LOG.error("Error while closing the websocket", e);
         }
         closeForwarder();
@@ -314,6 +356,21 @@ public class PortForwarderWebsocket implements PortForwarder {
       @Override
       public boolean isAlive() {
         return alive.get();
+      }
+
+      @Override
+      public boolean errorOccurred() {
+        return errorOccurred.get() || !clientThrowables.isEmpty() || !serverThrowables.isEmpty();
+      }
+
+      @Override
+      public Collection<Throwable> getClientThrowables() {
+        return clientThrowables;
+      }
+
+      @Override
+      public Collection<Throwable> getServerThrowables() {
+        return serverThrowables;
       }
     };
   }